requirements.txt
opencv-python
Pillow
transformers
torch
deepface
scikit-learn
flask
piexif
requests
tqdm
streamlit